# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hoshiyomi API",
    "description": "hoshiyomi 后端 REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "User__list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码,从 1 开始",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "每页条数,1-100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户分页列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PageData_UserResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "operationId": "User__create",
        "responses": {
          "200": {
            "description": "创建的用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "User__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "operationId": "User__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/api/v1/users/{id}/password": {
      "put": {
        "tags": [
          "user"
        ],
        "operationId": "User__changePassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "修改结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "无权修改他人密码或旧密码错误"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/api/v1/users/{id}/username": {
      "put": {
        "tags": [
          "user"
        ],
        "operationId": "User__updateUsername",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsernameReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新后的用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          },
          "409": {
            "description": "用户名已存在"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse_MessageResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
      "ApiResponse_PageData_UserResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "description": "分页数据载荷",
            "required": [
              "items",
              "total",
              "page",
              "per_page"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "id",
                    "username",
                    "email",
                    "created_at",
                    "updated_at"
                  ],
                  "properties": {
                    "created_at": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 0
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                }
              },
              "page": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "per_page": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "total": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
...
      "ApiResponse_UserResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "id",
              "username",
              "email",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "created_at": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "id": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "updated_at": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
...
      "ChangePasswordReq": {
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string"
          },
          "old_password": {
            "type": "string"
          }
        }
      },
      "FieldViolation": {
        "type": "object",
        "description": "某个字段的一条失败规则",
        "required": [
          "code",
          "params"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "失败的规则名(如 \"length\"、\"email\")"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "DTO 为该规则声明的消息(声明了才有)"
          },
          "params": {
            "type": "object",
            "description": "失败规则的参数(如 {\"min\": 3, \"max\": 20})。提交的字段值刻意不回显:validator 会把它记为 params.value,原样回显会把原始输入(口令、令牌)返回给客户端。",
            "additionalProperties": true
          }
        }
      },
...
      "UpdateUsernameReq": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
...
      "ValidationErrors": {
        "type": "object",
        "description": "按字段分组的全部违规。键为扁平化的字段路径:顶层字段是 email,嵌套结构体字段是 profile.email,列表元素是 items[0].name。映射有序,故序列化结果稳定。",
        "additionalProperties": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/FieldViolation"
          }
        },
        "propertyNames": {
          "type": "string"
        }
      }
    }
  }
}